@@ -13,7 +13,7 @@ import 'package:flutter_midi_command_example/main.dart';
 void main() {
   testWidgets('Verify Platform version', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(const MyApp());
+    await tester.pumpWidget(MyApp());
 
     // Verify that platform version is retrieved.
     expect(
